Remove StringExensions
