+++ src/WebJobs.Extensions.DurableTask/Microsoft.Azure.WebJobs.Extensions.DurableTask.Telemetry.xml
@@ -1562,18 +1562,6 @@
             Implements  the <see cref="M:Microsoft.ApplicationInsights.Channel.ITelemetryChannel.Flush" /> method.
             </summary>
         </member>
-        <member name="T:Microsoft.Azure.WebJobs.Extensions.DurableTask.Correlation.StringExtensions">
-            <summary>
-            StringExtensions are the extension method parse string into <see cref="T:Microsoft.Azure.WebJobs.Extensions.DurableTask.Correlation.TraceParent"></see> object.
-            </summary>
-        </member>
-        <member name="M:Microsoft.Azure.WebJobs.Extensions.DurableTask.Correlation.StringExtensions.ToTraceParent(System.String)">
-            <summary>
-            Convert a traceParent string to TraceParent object.
-            </summary>
-            <param name="traceparent">string representations of traceParent.</param>
-            <returns>TraceParent object.</returns>
-        </member>
         <member name="T:Microsoft.Azure.WebJobs.Extensions.DurableTask.Correlation.TelemetryActivator">
             <summary>
             TelemetryActivator activates Distributed Tracing. This class only works for netstandard2.0.
@@ -1639,6 +1627,13 @@
             <summary>
             Gets or sets the TraceFlags of the traceParent.
             </summary>
+        </member>
+        <member name="M:Microsoft.Azure.WebJobs.Extensions.DurableTask.Correlation.TraceParent.FromString(System.String)">
+            <summary>
+            Convert a traceParent string to TraceParent object.
+            </summary>
+            <param name="traceparent">string representations of traceParent.</param>
+            <returns>TraceParent object.</returns>
         </member>
         <member name="T:Microsoft.Azure.WebJobs.Extensions.DurableTask.DeterministicAttribute">
             <summary>
